pyproject.toml
[project]
name = "qa_autoautomation"
version = "0.1.0"
description = "LLM-powered pipeline that converts TestRail XML test cases into Playwright automated tests."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "pytest>=8.2.0",
    "playwright>=1.51.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
